Major refactoring of communication
As mentioned yesterday the communication was broken because assyncronous operations were being shipped that when deleted continued to use resources. Now all asyncronous operations capture a copy of the shared pointer. To achieve this behavior we add to change the constructors of the sockets clients and servers into a multistage construction process where its the static methods that return the shared pointers to these entities that trigger the asyncronous operations