Major refactoring of build system and upgrade of vcpkg
This commit tries to remove the stupid build arquitecture which requires to have acess to the internet to pull vsg, vsgimgui and vsg mesh reader. For us to do this we had to also upgrade vcpkg to a new version which included these ports. Sadly this upgrade broke some code related with rendering things on screen. We also had to fix this part of the code