New volumes for registration and new version of registration for dual sided cuts
USE THE CT POST LAPLACIAN and NOT THE RAW. In slicer they appear to be the same but the raw has noise with values very close to 0, so in slicer they do not show up well. However, these are identified as points when extracting the point clouds for PCA, causing the centroid to be different from the US one, which causes poor initial alignement and ITK to give exceptions because the images do not overlap enough.