This is a commit to properly deal with async resources for process management
So this is a tricky topic:
The process management requires pipes, but because there are pending async operations waiting on these resources, we cannot just delete them. We could explictly keep track of them and once all pending async operations have been finished we could submit new async operations, but this is tricky, and requires a lot of attention. Instead for each launching process we create a PendingAsyncData structure that must be a shared ptr. Internally this class contains all the resources of a given process. We submit this shared pointer to the asio context, when we want to delete said resource we can just loose our shared pointer and eventually all resources will be deleted from the async pipeline