Registration updates
Now the laplacian uses an histogram to define the threashold value, resgister_volumes_v9 is not ready yet, still has a lot of junk and unnecessary code, final_icp_registration just uses the icp method but it is finished (just the necessary code, well structrured and commented). Also, the preprocessing is now done in the registration .cpp itself, eliminating the need to apply the laplacian in separate. To use, input the raw side cutted volumes and ajust laplacian parameters if necessary.