IC_MI registration now features a safety check for outlier solution detection
The rotation matrices are extracted from both the solution of ICP and MI in the origin. The rotation matrix conrreponding to the transformation of one solution to another is computed and transformed into a angle-axis representation. If this angle is above a certain threashold, the MI solution is invalid and the next one with the next best optimizer value is evaluated until a valid solution is found.