This corrects the build system so that we can infer directly in C++ through pytorch
The main idea is that pytorch (for some obscure reason) cannot support static linkage, thus we need to add this as dynamic linkage, but if the whole project is linked dynamically, then imgui fails because it cannot be compiled with dynamic linkage. The solution is to specify by hand which libraries are compiled statically and which libraries are compiled dynamically so that the build does not fail. This is achieved through a custom triplet