dicom compliant testing
So the dicom viewer already cuts all points according to physical coordinates, which is good! now I need to write some stupid tests, just to make sure that the clipping is perfect for all geometries. Once this is done, off to the races with rendering the registration in real time. This will be tricky but I might b able to pull it off