Creating demo for hospital
Ok, so we need to create a custom demo that takes a CT and MRI image, sets the CT as baseline, performs registration on the MRI whilst the CT-MRI is viewed, simultaneously. For that we first add a new custom drawing call that is set when the optimizer is running.