MRI related information
ok so to add the MRI related information, we need to duplicate all maps that contain the volumes. This translates into also tracking the current volume that is being shown, so that when we transition from pages, we know from which volume alignment we are coming from and going to.