Update main.cpp
Update the medical viewing capabilities, now the only thing missing is to sample the image in real time. For that I will make a local copy of the moving image. At each iteration I take the moving image location, I move it,  then use a resampler, aligned with the fixed image, and then I render it.